+++ packages/igx-templates/Update.spec.ts
@@ -16,19 +16,19 @@ describe("updateWorkspace - Unit tests", () => {
 		fsSpy = jasmine.createSpyObj("fsSpy", ["fileExists", "directoryExists", "readFile", "writeFile", "glob"]);
 		spyOn(App.container, "get").and.returnValue(fsSpy);
 		spyOnProperty(App, "workDir", "get").and.returnValue("mockDir");
-		spyOn(ProjectConfig, "globalConfig").and.returnValue({});
+		spyOn(ProjectConfig, "getConfig").and.returnValue({});
 	});
 	it("Should fail if current used package is registry package", async () => {
 		spyOn(pkgResolve, "resolvePackage").and.returnValue(pkgResolve.FEED_PACKAGE);
-		expect(await updateWorkspace()).toEqual(false);
+		expect(await updateWorkspace("")).toEqual(false);
 	});
 
 	it("Should fail if '@infragistics' node modules are installed", async () => {
 		spyOn(pkgResolve, "resolvePackage").and.returnValue(pkgResolve.NPM_PACKAGE);
 		(fsSpy.directoryExists as jasmine.Spy).and.returnValue(true);
 		spyOn(Util, "log");
-		expect(await updateWorkspace()).toEqual(false);
-		expect(fsSpy.directoryExists).toHaveBeenCalledWith(path.join("mockDir", "node_modules", "@infragistics"));
+		expect(await updateWorkspace("root")).toEqual(false);
+		expect(fsSpy.directoryExists).toHaveBeenCalledWith(path.join("mockDir", "root", "node_modules", "@infragistics"));
 		expect(Util.log).toHaveBeenCalledWith("@infragistics module already exists. Nothing to do here.");
 	});
 
@@ -37,7 +37,7 @@ describe("updateWorkspace - Unit tests", () => {
 		(fsSpy.directoryExists as jasmine.Spy).and.returnValue(false);
 		(fsSpy.readFile as jasmine.Spy).and.returnValue("");
 		spyOn(Util, "log");
-		expect(await updateWorkspace()).toEqual(false);
+		expect(await updateWorkspace("")).toEqual(false);
 		expect(Util.log).toHaveBeenCalledWith("Package.json not found!");
 	});
 
@@ -49,13 +49,16 @@ describe("updateWorkspace - Unit tests", () => {
 		(fsSpy.directoryExists as jasmine.Spy).and.returnValue(false);
 		(fsSpy.readFile as jasmine.Spy).and.returnValue(JSON.stringify(mockJSONInput, null, 4));
 		spyOn(PackageManager, "ensureRegistryUser").and.returnValue(false);
-		expect(await updateWorkspace()).toEqual(false);
+		expect(await updateWorkspace("")).toEqual(false);
 	});
 
 	it("Should update package.json file, removing non-scoped igniteui-angular packages", async () => {
 		const mockPackageJSON = {
 			dependencies: {
+				"@alphabetically-sorted-scope/package": "^0.0.0",
+				"alphabetically-second-package": "^0.0.0",
 				"igniteui-angular": "^9.1.0",
+				"igniteui-dockmanager": "^1.0.0",
 				"some-package": "^0.0.0"
 			}
 		};
@@ -72,11 +75,14 @@ describe("updateWorkspace - Unit tests", () => {
 		});
 		spyOn(PackageManager, "ensureRegistryUser").and.returnValue(true);
 		spyOn(Util, "log");
-		expect(await updateWorkspace()).toEqual(false);
+		expect(await updateWorkspace("")).toEqual(false);
 		expect(fsSpy.writeFile).toHaveBeenCalledWith("package.json", JSON.stringify({
 			dependencies: {
-				"some-package": "^0.0.0",
-				"@infragistics/igniteui-angular": "^9.1.0"
+				"@alphabetically-sorted-scope/package": "^0.0.0",
+				"@infragistics/igniteui-angular": "^9.1.0",
+				"@infragistics/igniteui-dockmanager": "^1.0.0",
+				"alphabetically-second-package": "^0.0.0",
+				"some-package": "^0.0.0"
 			}
 		}, null, 2));
 	});
@@ -94,6 +100,7 @@ describe("updateWorkspace - Unit tests", () => {
 `import { something } from 'module';
 import { bait } from 'igniteui-angular-core';
 import { IgxGridComponent } from 'igniteui-angular';
+import { IgcDockManager } from 'igniteui-dockmanager';
 
 export class HomeComponent {
 	title = 'igniteui-angular example';
@@ -102,6 +109,7 @@ export class HomeComponent {
 `import { something } from 'module';
 import { bait } from 'igniteui-angular-core';
 import { IgxGridComponent } from '@infragistics/igniteui-angular';
+import { IgcDockManager } from '@infragistics/igniteui-dockmanager';
 
 export class HomeComponent {
 	title = 'igniteui-angular example';
@@ -110,11 +118,13 @@ export class HomeComponent {
 			content:
 `@import '~igniteui-angular/lib/core/styles/themes/utilities';
 @import '~igniteui-angular/lib/core/styles/themes/index';
+@import '~igniteui-dockmanager/styles/themes/test';
 @include igx-core();
 `,
 			expected:
 `@import '~@infragistics/igniteui-angular/lib/core/styles/themes/utilities';
 @import '~@infragistics/igniteui-angular/lib/core/styles/themes/index';
+@import '~@infragistics/igniteui-dockmanager/styles/themes/test';
 @include igx-core();
 `}, {
 			path: "angular.json",
@@ -127,7 +137,8 @@ export class HomeComponent {
 			"root": "",
 			"styles": [
 				"src/styles.scss",
-				"node_modules/igniteui-angular/styles/igniteui-angular.css"
+				"node_modules/igniteui-angular/styles/igniteui-angular.css",
+				"node_modules/igniteui-dockmanager/styles/themes/test"
 			]
 		}
 	}
@@ -141,12 +152,33 @@ export class HomeComponent {
 			"root": "",
 			"styles": [
 				"src/styles.scss",
-				"node_modules/@infragistics/igniteui-angular/styles/igniteui-angular.css"
+				"node_modules/@infragistics/igniteui-angular/styles/igniteui-angular.css",
+				"node_modules/@infragistics/igniteui-dockmanager/styles/themes/test"
 			]
 		}
 	}
+}`}, {
+	path: "src/app.module.ts",
+	content:
+`import { something } from 'module';
+import { bait } from 'igniteui-angular-core';
+import { IgxGridComponent } from 'igniteui-angular';
+import { dockManagerLoader } from 'igniteui-dockmanager/loader';
+
+export class HomeComponent {
+title = 'igniteui-angular example';
+}`,
+	expected:
+`import { something } from 'module';
+import { bait } from 'igniteui-angular-core';
+import { IgxGridComponent } from '@infragistics/igniteui-angular';
+import { dockManagerLoader } from '@infragistics/igniteui-dockmanager/loader';
+
+export class HomeComponent {
+title = 'igniteui-angular example';
 }`}];
-		(fsSpy.glob as jasmine.Spy).and.returnValue(["src/home.component.ts", "src/home.component.scss"]);
+		(fsSpy.glob as jasmine.Spy).and.returnValue(
+			["src/home.component.ts", "src/home.component.scss", "src/app.module.ts"]);
 		(fsSpy.readFile as jasmine.Spy).and.callFake((filePath: string) => {
 			if (filePath.indexOf("package.json") > -1) {
 				return JSON.stringify(mockPackageJSON);
@@ -156,7 +188,7 @@ export class HomeComponent {
 		});
 		(fsSpy.fileExists as jasmine.Spy).and.returnValue(true);
 		spyOn(PackageManager, "ensureRegistryUser").and.returnValue(true);
-		expect(await updateWorkspace()).toEqual(true);
+		expect(await updateWorkspace("")).toEqual(true);
 		for (const fileEntry of mockFileArray) {
 			expect(fsSpy.writeFile).toHaveBeenCalledWith(fileEntry.path, fileEntry.expected);
 		}