+++ spec/unit/ProjectConfig-spec.ts
@@ -4,19 +4,18 @@ import * as os from "os";
 
 describe("Unit - ProjectConfig", () => {
 	it("hasLocalConfig returns correct values", async done => {
-
 		const cwdSpy = spyOn(process, "cwd");
 		spyOn(os, "homedir").and.returnValues("rootDir");
-		spyOn(fs, "existsSync").and.returnValue(true);
+		spyOn(fs, "statSync").and.returnValue({ isFile: () => true });
 
 		// cwd matches homedir
 		cwdSpy.and.returnValue("rootDir");
 		expect(ProjectConfig.hasLocalConfig()).toBeFalsy();
-		expect(fs.existsSync).toHaveBeenCalledTimes(0);
+		expect(fs.statSync).toHaveBeenCalledTimes(0);
 
 		cwdSpy.and.returnValue("rootDir/somePath");
 		expect(ProjectConfig.hasLocalConfig()).toBeTruthy();
-		expect(fs.existsSync).toHaveBeenCalledTimes(1);
+		expect(fs.statSync).toHaveBeenCalledTimes(1);
 
 		done();
 	});